Slight modification to creation of config/injector.properties file
Now the parameters must be written in the config/injector.properties.template file.
Then the runInjector.py script will search and replace the number of peers to generate
the config/injector.properties file. This latter file is ignored by git, due to the use of
a local .gitignore in the config folder.